bring nbox.cli code to nbox.__main__
